Remove warnings.

+++ rtl-find-file.h
@@ -36,7 +36,7 @@ extern "C" {
 bool rtems_rtl_find_file (const char*  name,
                           const char*  paths,
                           const char** file_name,
-                          uint32_t*    size);
+                          size_t*      size);
 
 #ifdef __cplusplus
 }
